requirements.txt
aiogram>=3.13.0
tortoise-orm>=0.21.0
aerich>=0.7.2
python-dotenv>=1.0.1
matplotlib
APScheduler>=3.10.4,<4.0.0
pytz
